+++ src/Products/PloneMeeting/migrations/migrate_to_4208.py
@@ -1,5 +1,13 @@
 # -*- coding: utf-8 -*-
 
+from collective.documentgenerator.content.pod_template import IPODTemplate
+from collective.documentgenerator.content.style_template import IStyleTemplate
+from collective.documentviewer.settings import GlobalSettings
+from collective.iconifiedcategory.content.category import ICategory
+from collective.iconifiedcategory.content.subcategory import ISubcategory
+from plone.app.blob.interfaces import IATBlob
+from Products.CMFCore.Expression import Expression
+from Products.CMFPlone.utils import base_hasattr
 from Products.PloneMeeting.migrations import logger
 from Products.PloneMeeting.migrations import Migrator
 
@@ -24,6 +32,33 @@ class Migrate_To_4208(Migrator):
                 cfg.setUsedMeetingAttributes(mAttrs)
         logger.info('Done.')
 
+    def _fixDocumentviewerLayout(self):
+        """Now that auto_select_layout is False for documentviewer, make sure
+           AnnexTypes and PodTemplates created before do not use this layout."""
+        logger.info('Remove documentviewer layout for annex types and pod templates...')
+        # disable auto_select_layout
+        viewer_settings = GlobalSettings(self.portal)._metadata
+        viewer_settings['auto_select_layout'] = False
+        brains = self.catalog(object_provides=(
+            ICategory.__identifier__,
+            ISubcategory.__identifier__,
+            IPODTemplate.__identifier__,
+            IStyleTemplate.__identifier__,
+            IATBlob.__identifier__))
+        for brain in brains:
+            obj = brain.getObject()
+            if base_hasattr(obj, 'layout'):
+                obj.manage_delProperties(('layout', ))
+        logger.info('Done.')
+
+    def _updateAnnexDecisionDownloadAction(self):
+        """Update the annexDecision portal_type download action condition_expr."""
+        logger.info('Update annexDecision portal_type download action condition_expr...')
+        annexDecion = self.portal.portal_types['annexDecision']
+        action = annexDecion.getActionObject('object_buttons/download')
+        action.condition = Expression('python:object.show_download()')
+        logger.info('Done.')
+
     def run(self, extra_omitted=[], from_migration_to_4200=False):
 
         logger.info('Migrating to PloneMeeting 4208...')
@@ -39,7 +74,9 @@ class Migrate_To_4208(Migrator):
             # re-apply actions.xml to update documentation url
             self.ps.runImportStepFromProfile('profile-Products.PloneMeeting:default', 'actions')
             self._updateMeetingOptionalBooleanAttrs()
+            self._updateAnnexDecisionDownloadAction()
 
+        self._fixDocumentviewerLayout()
         logger.info('Migrating to PloneMeeting 4208... Done.')
 
 
@@ -47,7 +84,9 @@ def migrate(context):
     '''This migration function will:
 
        1) Update searches_decisions faceted config;
-       2) Re-apply actions.xml to update documentation URL.
+       2) Re-apply actions.xml to update documentation URL;
+       3) Remove documentviewer layout from annex types and pod templates;
+       4) Update annexDecision download action condition_expr.
     '''
     migrator = Migrate_To_4208(context)
     migrator.run()